# docs/swagger.yaml
basePath: /
definitions:
...
  models.AllTickets:
    properties:
      closed_tickets:
        $ref: '#/definitions/models.ClosedTickets'
      open_tickets:
        $ref: '#/definitions/models.OpenTickets'
      pending_tickets:
        $ref: '#/definitions/models.PendingTickets'
      resolved_tickets:
        $ref: '#/definitions/models.ResolvedTickets'
    type: object
...
  models.ClosedTickets:
    properties:
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      total:
        type: integer
    type: object
...
  models.OpenTickets:
    properties:
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      total:
        type: integer
    type: object
  models.PendingTickets:
    properties:
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      total:
        type: integer
    type: object
...
  models.Requester:
    properties:
      email:
        type: string
      id:
        type: integer
      mobile:
        type: integer
      name:
        type: string
      phone:
        type: string
    type: object
  models.ResolvedTickets:
    properties:
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      total:
        type: integer
    type: object
...
  models.Stats:
    properties:
      agent_responded_at: {}
      closed_at: {}
      first_responded_at: {}
      pending_since: {}
      reopened_at: {}
      requester_responded_at: {}
      resolved_at: {}
      status_updated_at:
        type: string
    type: object
  models.Ticket:
    properties:
      associated_tickets_count: {}
      attachments:
        items: {}
        type: array
      category: {}
      cc_emails:
        items: {}
        type: array
      created_at:
        type: string
      custom_fields:
        type: object
      deleted:
        type: boolean
      department_id: {}
      description:
        type: string
      description_text:
        type: string
      due_by:
        type: string
      email_config_id: {}
      fr_due_by:
        type: string
      fr_escalated:
        type: boolean
      fwd_emails:
        items: {}
        type: array
      group_id: {}
      id:
        type: integer
      is_escalated:
        type: boolean
      item_category: {}
      nr_due_by: {}
      nr_escalated:
        type: boolean
      priority: {}
      reply_cc_emails:
        items: {}
        type: array
      requested_for_id:
        type: integer
      requester:
        $ref: '#/definitions/models.Requester'
      requester_id:
        type: integer
      responder_id: {}
      source: {}
      source_additional_info: {}
      spam:
        type: boolean
      stats:
        $ref: '#/definitions/models.Stats'
      status: {}
      sub_category: {}
      subject:
        type: string
      tags:
        items:
          type: string
        type: array
      to_emails: {}
      twitter_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8012
info:
  contact:
    email: eti@cisco.io
    name: API Support
  description: This is a Kosha REST serice for exposing many freshservice features
    as REST APIs with better consistency, observability etc
  termsOfService: http://swagger.io/terms/
  title: Freshservice Connector API
  version: "2.0"
paths:
...
  /api/v2/stats/agent/{agent_id}:
    get:
      consumes:
      - application/json
      description: Get ticket statistics for a particular agent id
      parameters:
      - description: Enter agent id
        in: path
        name: agent_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllTickets'
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get ticket statistics for a particular agent
      tags:
      - statistics
  /api/v2/stats/date/{created_at}:
    get:
      consumes:
      - application/json
      description: Get ticket statistics for tickets created since a particular date
      parameters:
      - description: Enter created_at
        in: path
        name: created_at
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllTickets'
      summary: Get ticket statistics for tickets created since a particular date
      tags:
      - statistics
  /api/v2/stats/group/{group_id}:
    get:
      consumes:
      - application/json
      description: Get ticket statistics for a particular group id
      parameters:
      - description: Enter group id
        in: path
        name: group_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllTickets'
      summary: Get ticket statistics for a particular group
      tags:
      - statistics
  /api/v2/tickets:
    get:
      consumes:
      - application/json
      description: List all tickets logged in the system
      parameters:
      - description: Page number
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ticket'
            type: array
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get all tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Create single ticket in the system
      parameters:
      - description: Enter ticket properties
        in: body
        name: text
        schema:
          $ref: '#/definitions/models.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create new ticket
      tags:
      - tickets
  /api/v2/tickets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete single ticket based on the ticket id logged in the system
      parameters:
      - description: Enter ticket id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete single ticket
      tags:
      - tickets
    get:
      consumes:
      - application/json
      description: List single ticket based on the ticket id logged in the system
      parameters:
      - description: Enter ticket id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get single ticket
      tags:
      - tickets
swagger: "2.0"